+++ project-addons/pmp_landed_costs/i18n/es.po
@@ -124,3 +124,106 @@ msgstr "No todos los productos tienen HS Code: %s"
 #, python-format
 msgid "Not all the products have weight: %s"
 msgstr "No todos los productos tienen peso: %s"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_stock_container_container_type
+#: model:ir.model.fields,help:pmp_landed_costs.field_stock_container_container_type
+msgid "Container type"
+msgstr "Tipo de contenedor"
+
+#. module: pmp_landed_costs
+#: model:ir.actions.act_window,name:pmp_landed_costs.action_stock_container_import_sheets
+msgid "Stock Container Import Sheets"
+msgstr "Hojas de Importación"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_dua_date
+msgid "DUA date"
+msgstr "Fecha DUA"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_forwarder_comercial
+msgid "Forwarded"
+msgstr "Transitario"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_treasury
+msgid "Treasury"
+msgstr "Hacienda"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_freight
+msgid "Freight"
+msgstr "Flete"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_fee
+msgid "Fee"
+msgstr "Arancel"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_inspection
+msgid "Inspection"
+msgstr "Inspección"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_arrival_cost
+msgid "Arrival costs"
+msgstr "Costes de llegada"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_container_type
+msgid "Container Type"
+msgstr "Contenedor"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_channel
+msgid "Channel"
+msgstr "Canal"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_stock_container_form
+#: model:ir.ui.menu,name:pmp_landed_costs.import_sheet_menu
+#: model:ir.actions.act_window,name:pmp_landed_costs.action_open_import_sheet_tree_view
+msgid "Import Sheets"
+msgstr "Hojas de Importación"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_stock_container_form
+msgid "Create New Sheet"
+msgstr "Crear nueva hoja"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_stock_container_add_import_sheets_form
+msgid "Importations"
+msgstr "Importaciones"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_import_sheet_form
+msgid "General"
+msgstr "General"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_import_sheet_form
+msgid "Characteristics"
+msgstr "Características"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_import_sheet_form
+msgid "Costs"
+msgstr "Costes"
+
+#. module: pmp_landed_costs
+#: selection:purchase.order,invoice_status:0
+msgid "Partially Invoiced"
+msgstr "Parcialmente facturado"
+
+#. module: pmp_landed_costs
+#: model:ir.actions.report,name:pmp_landed_costs.import_sheet_xlsx
+msgid "Print to XLSX"
+msgstr "Imprimir a XLSX"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_container_id
+msgid "Container"
+msgstr "Envío"
