+++ project-addons/pmp_landed_costs/i18n/it.po
@@ -123,3 +123,106 @@ msgstr "Non tutti i prodotti hanno il codice HS: %s"
 #, python-format
 msgid "Not all the products have weight: %s"
 msgstr "Non tutti i prodotti hanno il peso: %s"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_container_id
+msgid "Container"
+msgstr "Spedizione"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_stock_container_container_type
+#: model:ir.model.fields,help:pmp_landed_costs.field_stock_container_container_type
+msgid "Container type"
+msgstr "Tipo di contenitore"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_stock_container_container_type
+#: model:ir.model.fields,help:pmp_landed_costs.field_stock_container_channel
+msgid "Channel"
+msgstr "Canale"
+
+#: model:ir.actions.act_window,name:pmp_landed_costs.action_stock_container_import_sheets
+msgid "Stock Container Import Sheets"
+msgstr "Fogli di importazioni"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_dua_date
+msgid "DUA date"
+msgstr "Data DUA"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_forwarder_comercial
+msgid "Forwarded"
+msgstr "Spedizioniere"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_treasury
+msgid "Treasury"
+msgstr "Tesoro"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_freight
+msgid "Freight"
+msgstr "Transporto merci"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_fee
+msgid "Fee"
+msgstr "Tariffa"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_inspection
+msgid "Inspection"
+msgstr "Ispezione"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_arrival_cost
+msgid "Arrival costs"
+msgstr "Costi d'arrivo"
+
+#. module: pmp_landed_costs
+#: model:ir.model.fields,field_description:pmp_landed_costs.field_import_sheet_container_type
+msgid "Container Type"
+msgstr "Contenitore"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_stock_container_form
+#: model:ir.ui.menu,name:pmp_landed_costs.import_sheet_menu
+#: model:ir.actions.act_window,name:pmp_landed_costs.action_open_import_sheet_tree_view
+msgid "Import Sheets"
+msgstr "Fogli di importazioni"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_stock_container_form
+msgid "Create New Sheet"
+msgstr "Crea nuovo foglio"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_stock_container_add_import_sheets_form
+msgid "Importations"
+msgstr "Importazioni"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_import_sheet_form
+msgid "General"
+msgstr "Generale"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_import_sheet_form
+msgid "Characteristics"
+msgstr "Caratteristiche"
+
+#. module: pmp_landed_costs
+#: model:ir.ui.view,arch_db:pmp_landed_costs.view_import_sheet_form
+msgid "Costs"
+msgstr "Costi"
+
+#. module: pmp_landed_costs
+#: selection:purchase.order,invoice_status:0
+msgid "Partially Invoiced"
+msgstr "Parzialmente Fatturato"
+
+#. module: pmp_landed_costs
+#: model:ir.actions.report,name:pmp_landed_costs.import_sheet_xlsx
+msgid "Print to XLSX"
+msgstr "Imprimere un XLSX"
