+++ project-addons/pmp_landed_costs/models/__init__.py
@@ -1,5 +1,4 @@
 # Copyright 2019 Omar Castiñeira, Comunitea Servicios Tecnológicos S.L.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import product
-from . import stock_landed_costs
+from . import product, stock_landed_costs, stock, import_sheet
