+++ project-addons/pmp_landed_costs/views/stock_view.xml
@@ -1,0 +1,33 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <record id="view_stock_container_add_import_sheets_form" model="ir.ui.view">
+        <field name="name">stock.container.add.import.sheets.form</field>
+        <field name="model">stock.container</field>
+        <field name="inherit_id" ref="purchase_picking.view_stock_container_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//notebook" position="inside">
+                <page string="Importations">
+                    <field name="import_sheet_ids" widget="one2many_list" colspan="4">
+                        <tree string="Import Sheets" class="custom_tree" default_order="dua asc">
+                            <field name="dua"/>
+                            <field name="container_id"/>
+                            <field name="dua_date"/>
+                            <field name="treasury"/>
+                            <field name="kgs"/>
+                            <field name="cbm"/>
+                            <field name="incoterm"/>
+                            <field name="destination_port"/>
+                            <field name="forwarder_comercial"/>
+                            <field name="container_type"/>
+                            <field name="channel"/>
+                            <field name="freight"/>
+                            <field name="fee"/>
+                            <field name="inspection"/>
+                            <field name="arrival_cost"/>
+                        </tree>
+                    </field>
+                </page>
+            </xpath>
+        </field>
+    </record>
+</odoo>
