+++ project-addons/purchase_picking/__manifest__.py
@@ -34,7 +34,6 @@
     "data": ['wizard/create_picking_move_view.xml',
              'views/purchase_view.xml',
              'data/res_partner_data.xml',
-             'views/import_sheet_view.xml',
              'views/stock_view.xml',
              'security/ir.model.access.csv',
              'security/purchase_picking_security.xml',
