+++ project-addons/purchase_picking/i18n/es.po
@@ -48,7 +48,6 @@ msgstr "Consumido para"
 #: model:ir.ui.view,arch_db:purchase_picking.stock_move_search_add_new_state
 #: model:ir.model.fields,field_description:purchase_picking.field_stock_move_container_id
 #: model:ir.model.fields,field_description:purchase_picking.field_stock_reservation_container_id
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_container_id
 msgid "Container"
 msgstr "Envío"
 
@@ -621,101 +620,3 @@ msgstr "Referencia Proveedor"
 #: model:ir.model.fields,field_description:purchase_picking.field_stock_container_customs_channel
 msgid "Customs Channel"
 msgstr "Canal"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_stock_container_container_type
-#: model:ir.model.fields,help:purchase_picking.field_stock_container_container_type
-msgid "Container type"
-msgstr "Tipo de contenedor"
-
-#. module: purchase_picking
-#: model:ir.actions.act_window,name:purchase_picking.action_stock_container_import_sheets
-msgid "Stock Container Import Sheets"
-msgstr "Hojas de Importación"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_dua_date
-msgid "DUA date"
-msgstr "Fecha DUA"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_forwarder_comercial
-msgid "Forwarded"
-msgstr "Transitario"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_treasury
-msgid "Treasury"
-msgstr "Hacienda"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_freight
-msgid "Freight"
-msgstr "Flete"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_fee
-msgid "Fee"
-msgstr "Arancel"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_inspection
-msgid "Inspection"
-msgstr "Inspección"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_arrival_cost
-msgid "Arrival costs"
-msgstr "Costes de llegada"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_container_type
-msgid "Container Type"
-msgstr "Contenedor"
-
-#. module: purchase_picking
-#: model:ir.model.fields,field_description:purchase_picking.field_import_sheet_channel
-msgid "Channel"
-msgstr "Canal"
-
-#. module: purchase_picking
-#: model:ir.ui.view,arch_db:purchase_picking.view_stock_container_form
-#: model:ir.ui.menu,name:purchase_picking.import_sheet_menu
-#: model:ir.actions.act_window,name:purchase_picking.action_open_import_sheet_tree_view
-msgid "Import Sheets"
-msgstr "Hojas de Importación"
-
-#. module: purchase_picking
-#: model:ir.ui.view,arch_db:purchase_picking.view_stock_container_form
-msgid "Create New Sheet"
-msgstr "Crear nueva hoja"
-
-#. module: purchase_picking
-#: model:ir.ui.view,arch_db:purchase_picking.view_stock_container_form
-msgid "Importations"
-msgstr "Importaciones"
-
-#. module: purchase_picking
-#: model:ir.ui.view,arch_db:purchase_picking.view_import_sheet_form
-msgid "General"
-msgstr "General"
-
-#. module: purchase_picking
-#: model:ir.ui.view,arch_db:purchase_picking.view_import_sheet_form
-msgid "Characteristics"
-msgstr "Características"
-
-#. module: purchase_picking
-#: model:ir.ui.view,arch_db:purchase_picking.view_import_sheet_form
-msgid "Costs"
-msgstr "Costes"
-
-#. module: purchase_picking
-#: selection:purchase.order,invoice_status:0
-msgid "Partially Invoiced"
-msgstr "Parcialmente facturado"
-
-#. module: purchase_picking
-#: model:ir.actions.report,name:purchase_picking.import_sheet_xlsx
-msgid "Print to XLSX"
-msgstr "Imprimir a XLSX"
