+++ project-addons/purchase_picking/models/__init__.py
@@ -1,1 +1,1 @@
-from . import stock, purchase, import_sheet
+from . import stock, purchase
